mod: move SingleJobScheduler into job/SingleJobScheduler.hpp
SingleJobScheduler is an IJobScheduler implementation, a sibling of the
interface (job/IJobScheduler.hpp) and MultiJobScheduler (job/MultiJobScheduler.hpp),
yet it lived in root-level JobRunner.hpp whose name only reflects the
back-compat alias. Move the class (and its ProcessingSync helper) into
job/SingleJobScheduler.hpp next to its siblings; JobRunner.hpp becomes a
thin compatibility header that includes it and keeps the
using JobRunner = SingleJobScheduler alias so existing includers are
unaffected.
